refactor: factories esperam receber dto no parametro do metodo
