+++ src/clientes/domain/factory/cliente.factory.ts
@@ -1,8 +1,17 @@
-import { Cliente } from '../cliente';
-import { CriarClienteDto } from '../../presenter/http/dto/criar-cliente.dto';
+import { Cliente } from '../cliente.entity';
+import { CriarClienteDto } from '../../presenter/http/dto/criarCliente.dto';
+import { randomUUID } from 'crypto';
 
 export class ClienteFactory {
   public criar(clienteDto: CriarClienteDto): Cliente {
-    return new Cliente(clienteDto.nome, clienteDto.cpf, clienteDto.email, clienteDto.telefone, clienteDto.cep);
+    return new Cliente(
+      randomUUID(),
+      clienteDto.nome,
+      clienteDto.cpf,
+      clienteDto.email,
+      clienteDto.telefone,
+      clienteDto.cep,
+      clienteDto.registroGerente,
+    );
   }
 }
